refactor: apply four_pillar srp, clean architecture
