+++ src/four_pillars/domain/constants.py
@@ -1,0 +1,24 @@
+"""사주 계산에 사용되는 상수"""
+
+# 천간 (天干)
+JIKKAN = ["甲", "乙", "丙", "丁", "戊", "己", "庚", "辛", "壬", "癸"]
+
+# 지지 (地支)
+JYUNISHI = [
+    "子",
+    "丑",
+    "寅",
+    "卯",
+    "辰",
+    "巳",
+    "午",
+    "未",
+    "申",
+    "酉",
+    "戌",
+    "亥",
+]
+
+# 기준 연도 (1864년)
+BASE_YEAR = 1864
+
