+++ src/four_pillars/domain/entities/enums.py
